refactored sidebar list
