+++ packages/databyss-ui/components/Sidebar/Sidebar.js
@@ -1,12 +1,5 @@
 import React, { useState } from 'react'
-import SourceSvg from '@databyss-org/ui/assets/source.svg'
-import AuthorSvg from '@databyss-org/ui/assets/author.svg'
-import PageSvg from '@databyss-org/ui/assets/page.svg'
-import TopicSvg from '@databyss-org/ui/assets/topic.svg'
-import Databyss from '@databyss-org/ui/assets/databyss.svg'
-import ArrowLeft from '@databyss-org/ui/assets/arrowLeft.svg'
-import ArrowRight from '@databyss-org/ui/assets/arrowRight.svg'
-
+import SidebarContent from './SidebarList'
 import {
   Text,
   View,
@@ -74,110 +67,6 @@ const Sidebar = ({ children }) => {
   const [menuOpen, setMenuOpen] = useState(true)
   const [menuItem, setMenuItem] = useState(false)
 
-  const composeList = type => {
-    console.log(type)
-  }
-
-  const menuItems = [
-    {
-      svg: menuOpen ? <ArrowLeft /> : <Databyss />,
-      text: 'Databyss',
-      action: () => setMenuOpen(!menuOpen),
-    },
-    {
-      svg: <PageSvg />,
-      text: 'Pages',
-      action: () => setMenuItem('Pages'),
-      list: [
-        {
-          svg: <ArrowLeft />,
-          text: 'Pages',
-          action: () => {
-            setMenuItem(false)
-          },
-        },
-        {
-          svg: <PageSvg />,
-          text: 'page title 1',
-          action: () => console.log('dispatch id 1'),
-        },
-        {
-          svg: <PageSvg />,
-          text: 'page title 2',
-          action: () => console.log('dispatch id 2'),
-        },
-        {
-          svg: <PageSvg />,
-          text: 'page title 3',
-          action: () => console.log('dispatch id 3'),
-        },
-      ],
-    },
-    {
-      svg: <SourceSvg />,
-      text: 'Sources',
-      action: () => setMenuItem('Sources'),
-    },
-    {
-      svg: <AuthorSvg />,
-      text: 'Authors',
-      action: () => setMenuItem('Authors'),
-    },
-    {
-      svg: <TopicSvg />,
-      text: 'Topics',
-      action: () => setMenuItem('Topics'),
-    },
-  ]
-
-  const SidebarContent = list =>
-    list.reduce((acc, item, index) => {
-      if (index !== 0) {
-        acc.push(<Separator color="border.1" key={`separator${index}`} />)
-      }
-      acc.push(
-        <BaseControl
-          key={index}
-          width="100%"
-          onClick={item.action}
-          alignItems={!menuOpen && 'flex-end'}
-        >
-          {menuOpen ? (
-            <View>
-              <Grid singleRow alignItems="center" columnGap="small">
-                <Icon sizeVariant={index ? 'tiny' : 'medium'} color="text.3">
-                  {item.svg}
-                </Icon>
-                <Text
-                  variant={index ? 'uiTextSmall' : 'uiTextLarge'}
-                  color="text.2"
-                >
-                  {item.text}
-                </Text>
-                {index && !menuItem ? (
-                  <View position="absolute" right="small">
-                    <Icon sizeVariant="small" color="text.3">
-                      <ArrowRight />
-                    </Icon>
-                  </View>
-                ) : null}
-              </Grid>
-            </View>
-          ) : (
-            <Grid singleRow id="here" alignItems="flex-end" columnGap="small">
-              <Icon
-                sizeVariant={!index || !menuOpen ? 'medium' : 'tiny'}
-                color="text.3"
-              >
-                {item.svg}
-              </Icon>
-            </Grid>
-          )}
-        </BaseControl>
-      )
-      return acc
-    }, [])
-
   return (
     <View alignItems="stretch" flexGrow={1} width="100%">
       <Grid columnGap="none" rowGap="none">
@@ -203,9 +92,14 @@ const Sidebar = ({ children }) => {
               mb="none"
               alignItems={menuOpen ? 'center' : 'flex-end'}
             >
-              {menuItem
-                ? SidebarContent(menuItems.find(i => i.text === menuItem).list)
-                : SidebarContent(menuItems)}
+              {/*
+            if menuItem exists, load list of items and compose 
+            */}
+              {menuItem ? (
+                <div>test</div>
+              ) : (
+                SidebarContent({ menuOpen, setMenuOpen, setMenuItem })
+              )}
             </List>
             {menuOpen && (
               <View position="fixed" bottom={0} left={0} width="100%">
